requirements.txt
# Your requirements go here
cachetools==5.5.0
google-auth==2.36.0
google-auth-oauthlib==1.2.1
gspread==6.1.4
oauthlib==3.2.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
requests-oauthlib==2.0.0
rsa==4.9